Create balance weights.c
In this example, there are 5 boxcars with different weights summing to 110.0. The output shows that we are modifying all the boxcars so that they each carry a weight of 22.0 (which makes a total of 110.0 for the entire train). So we remove 18.0 for the first boxcar, we add 10.0 for the second, we add 2.0 for the third